tested C code to write values to Hyve, commented it out for the moment

+++ Hyve/DashboardViewController.m
@@ -242,7 +242,7 @@
     }
 
 
-    NSLog(@"self.peripheralMutableArray.count %i", self.peripheralMutableArray.count);
+    NSLog(@"self.peripheralMutableArray.count %lu", (unsigned long)self.peripheralMutableArray.count);
 
 //    if ([peripheral.name isEqualToString:@"VLT’s MacBook Pro"])
 //    {
